fix: refactor codes into different folders
