+++ src/js/dom/closemodal.js
