+++ src/js/dom/todoRowInnerHtml.js
