+++ src/js/functions/additems.js
@@ -1,5 +1,5 @@
-import getItems from './getitems';
-import setItemToDatabase from './setItem';
+import getItems from '../storage/getitems';
+import setItemToDatabase from '../storage/setItem';
 
 const addItem = (item, databaseName) => {
   const items = getItems(databaseName);
