+++ src/js/functions/dynamicvalues.js
