--- src/js/setItem.js
@@ -1,5 +1,0 @@
-const setItemToDatabase = (databaseName,items) =>{
- localStorage.setItem(databaseName, JSON.stringify(items));
-}
-
-export default setItemToDatabase;
