+++ src/js/storage/setItem.js
@@ -1,0 +1,5 @@
+const setItemToDatabase = (databaseName, items) => {
+  localStorage.setItem(databaseName, JSON.stringify(items));
+};
+
+export default setItemToDatabase;
